fix: move candidate packages to utils folder
