+++ package.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/telicent-oss/telicent-ds.git"
   },
   "type": "module",
-  "version": "2.1.0",
+  "version": "2.1.1-0",
   "private": false,
   "dependencies": {
     "@emotion/react": "^11.10.6",
