+++ src/utils/split-uri-base-and-term/__tests__/splitUriBaseAndTerm.test.ts
