+++ src/utils/split-uri-base-and-term/index.ts
