+++ src/utils/utils-lib/src/GraphApp/constants.ts
